agent_architecture/pyproject.toml
[project]
name = "agent-architecture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.56.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
